+++ pkg/parser/parser.go
@@ -422,7 +422,8 @@ func IsNameChar(r byte) bool {
 		r == '?' || r == ':' ||
 		r == '[' || r == ']' ||
 		r == '^' || r == '$' ||
-		r == '<' || r == '>'
+		r == '<' || r == '>' ||
+		r == '&'
 }
 
 func isDigit(r byte) bool {