+++ CHANGES.md
@@ -15,6 +15,7 @@ CHANGELOG
 ---------
 **Master**
  - [Fix] `lineWidth` param is not working
+ - [Fix] Support `&` in metric name
  
 **0.9.2**
  - [Improvement] "pipe" function chaining syntax. Same way as in https://github.com/graphite-project/graphite-web/pull/2042 (thx. to @lomik)